fix: 修复 UBSAN 整数溢出错误，启用严格模式
- 将 test_full_flow.cpp 中的 volatile int 改为 volatile long long，
  避免 10000 次 i*i 累加导致的整数溢出
- 将 UBSAN_OPTIONS 的 halt_on_error 从 0 改为 1，
  使 UBSAN 检测到未定义行为时直接失败

+++ tests/test_full_flow.cpp
@@ -17,7 +17,7 @@ TEST(FullFlowTest, GperftoolsGeneratesValidProfile) {
 
     // 运行足够的工作负载
     std::cout << "Running workload...\n";
-    volatile int result = 0;
+    volatile long long result = 0;
     for (int i = 0; i < 10000; ++i) {
         result += i * i;
         // 一些函数调用
